add submodules support
